# internal/openapi/openapi.json
{
  "components": {
    "responses": {
...
      "NotAuthenticated": {
        "content": {
          "application/json": {
            "example": {
              "error": {
                "code": "not_authenticated"
              }
            },
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Authentication is required."
      },
...
      "OperationEnvironmentConflict": {
        "content": {
          "application/json": {
            "example": {
              "error": {
                "code": "environment_inactive",
                "message": "Activate Live before viewing its events."
              }
            },
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "The selected project environment is not active."
      },
      "OperationEnvironmentNotFound": {
        "content": {
          "application/json": {
            "example": {
              "error": {
                "code": "project_environment_not_found",
                "message": "This project environment is not available."
              }
            },
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "The selected project environment is unavailable."
      },
      "OperationListFailed": {
        "content": {
          "application/json": {
            "example": {
              "error": {
                "code": "operation_list_failed",
                "message": "Consumel could not load events. Try again shortly."
              }
            },
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "The event list could not be loaded."
      },
      "OperationListInvalid": {
        "content": {
          "application/json": {
            "examples": {
              "invalid_cursor": {
                "value": {
                  "error": {
                    "code": "invalid_cursor"
                  }
                }
              },
              "invalid_filter": {
                "value": {
                  "error": {
                    "code": "invalid_filter"
                  }
                }
              },
              "invalid_request": {
                "value": {
                  "error": {
                    "code": "invalid_request"
                  }
                }
              }
            },
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "The event list filter or pagination input is invalid."
      },
...
    },
    "schemas": {
...
      "Error": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "additionalProperties": false,
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code"
            ],
            "type": "object"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "UsageOperation": {
        "additionalProperties": false,
        "properties": {
          "balanceDebited": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "billable": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "customerId": {
            "type": "string"
          },
          "denialReason": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "lastReplayedAt": {
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          },
          "meterKey": {
            "type": "string"
          },
          "meterType": {
            "enum": [
              "prepaid",
              "postpaid",
              "hybrid"
            ],
            "type": "string"
          },
          "quantity": {
            "format": "int64",
            "minimum": 1,
            "type": "integer"
          },
          "remainingBalance": {
            "format": "int64",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          },
          "replayCount": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "status": {
            "enum": [
              "accepted",
              "denied"
            ],
            "type": "string"
          }
        },
        "required": [
          "id",
          "customerId",
          "meterKey",
          "quantity",
          "meterType",
          "status",
          "denialReason",
          "balanceDebited",
          "remainingBalance",
          "billable",
          "replayCount",
          "lastReplayedAt",
          "createdAt"
        ],
        "type": "object"
      },
      "UsageOperations": {
        "additionalProperties": false,
        "properties": {
          "nextCursor": {
            "type": [
              "string",
              "null"
            ]
          },
          "operations": {
            "items": {
              "$ref": "#/components/schemas/UsageOperation"
            },
            "type": "array"
          }
        },
        "required": [
          "operations",
          "nextCursor"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "localCookieSession": {
        "in": "cookie",
        "name": "consumel_session",
        "type": "apiKey"
      },
      "productionCookieSession": {
        "in": "cookie",
        "name": "__Host-consumel_session",
        "type": "apiKey"
      },
      "projectAPIKey": {
        "bearerFormat": "cm_test_… or cm_live_…",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Consumel usage-based billing infrastructure API.",
    "title": "Consumel API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/projects/{projectId}/environments/{environment}/events": {
      "get": {
        "parameters": [
          {
            "description": "The immutable ID of the project selected in the dashboard.",
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "The selected isolated project environment.",
            "in": "path",
            "name": "environment",
            "required": true,
            "schema": {
              "enum": [
                "sandbox",
                "live"
              ],
              "type": "string"
            }
          },
          {
            "description": "Filter by the persisted operation outcome.",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "enum": [
                "accepted",
                "denied"
              ],
              "type": "string"
            }
          },
          {
            "description": "Inclusive RFC 3339 start time. Supply from and to together; the range may span at most one year.",
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "description": "Exclusive RFC 3339 end time. Supply from and to together.",
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "description": "The opaque next cursor from the previous page.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "The number of events to return.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "nextCursor": null,
                  "operations": [
                    {
                      "balanceDebited": 500,
                      "billable": true,
                      "createdAt": "2026-09-29T12:00:00Z",
                      "customerId": "customer_123",
                      "denialReason": null,
                      "id": "0199aad1-f00d-7ae0-935f-cde0bd9f3ba5",
                      "lastReplayedAt": "2026-09-29T12:01:00Z",
                      "meterKey": "api_calls",
                      "meterType": "prepaid",
                      "quantity": 500,
                      "remainingBalance": 9500,
                      "replayCount": 1,
                      "status": "accepted"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/UsageOperations"
                }
              }
            },
            "description": "Completed successfully."
          },
          "400": {
            "$ref": "#/components/responses/OperationListInvalid"
          },
          "401": {
            "$ref": "#/components/responses/NotAuthenticated"
          },
          "404": {
            "$ref": "#/components/responses/OperationEnvironmentNotFound"
          },
          "409": {
            "$ref": "#/components/responses/OperationEnvironmentConflict"
          },
          "500": {
            "$ref": "#/components/responses/OperationListFailed"
          }
        },
        "security": [
          {
            "productionCookieSession": []
          },
          {
            "localCookieSession": []
          }
        ],
        "summary": "List accepted and denied usage operations in the signed-in project workspace.",
        "tags": [
          "Dashboard Events"
        ]
      }
    },
    "/v1/projects/{projectId}/environments/{environment}/events/stream": {
      "get": {
        "parameters": [
          {
            "description": "The immutable ID of the project selected in the dashboard.",
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "The selected isolated project environment.",
            "in": "path",
            "name": "environment",
            "required": true,
            "schema": {
              "enum": [
                "sandbox",
                "live"
              ],
              "type": "string"
            }
          },
          {
            "description": "The last Redis stream cursor received by the client. Browsers normally reconnect through Last-Event-ID automatically.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "pattern": "^[0-9]+-[0-9]+$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "A server-sent event stream. Each usage.operation event contains a UsageOperation JSON payload and a Redis stream cursor used for reconnection."
          },
          "400": {
            "$ref": "#/components/responses/OperationListInvalid"
          },
          "401": {
            "$ref": "#/components/responses/NotAuthenticated"
          },
          "404": {
            "$ref": "#/components/responses/OperationEnvironmentNotFound"
          },
          "409": {
            "$ref": "#/components/responses/OperationEnvironmentConflict"
          },
          "500": {
            "$ref": "#/components/responses/OperationListFailed"
          }
        },
        "security": [
          {
            "productionCookieSession": []
          },
          {
            "localCookieSession": []
          }
        ],
        "summary": "Stream newly persisted usage operations to the signed-in project workspace.",
        "tags": [
          "Dashboard Events"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Production",
      "url": "https://api.consumel.com"
    },
    {
      "description": "Local development",
      "url": "http://localhost:8080"
    }
  ]
}